+++ packages/core/src/contexts/DriverContext.ts
@@ -13,9 +13,6 @@ export default class DriverContext<T = DriverArgs> extends ConfigContext<T> {
   // Name defined on the plugin (kebab case)
   driverName: string = '';
 
-  // Name used for emitting events (kebab case)
-  eventName: string;
-
   // Nested list of argv for each parallel execution
   parallelArgv: Argv[] = [];
 
@@ -26,7 +23,6 @@ export default class DriverContext<T = DriverArgs> extends ConfigContext<T> {
     super(args);
 
     this.driverName = driver.name;
-    this.eventName = driver.name;
     this.parallelArgv = parallelArgv;
     this.primaryDriver = driver;