+++ packages/core/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 #### 💥 Breaking
 
+- Migrated to the new [@boost/event](https://milesj.gitbook.io/boost/event) system.
 - Script names (on the command line) will now error if not in kebab case.
 - Renamed `Context#root` to `cwd`.
+- Removed `DriverContext#eventName` and `ScriptContext#eventName`.
 - Removed `ScriptContext#binName`. Use `scriptName` instead.
 
 #### 🚀 Updates
@@ -25,5 +27,6 @@
 
 #### 🛠 Internals
 
+- `Beemo` now extends from `Tool` instead of managing an instance.
 - Updated `hygen` to v4.
 - Updated dependencies.